requirements.txt
altair==4.2.0
antlr4-python3-runtime==4.9.3
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.5
braceexpand==0.1.7
cachetools==5.2.0
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
clip==1.0
commonmark==0.9.1
contexttimer==0.3.3
cycler==0.11.0
decorator==5.1.1
decord==0.6.0
distlib==0.3.6
e==1.4.5
einops==0.5.0
entrypoints==0.4
fairscale==0.4.4
filelock==3.8.0
fonttools==4.37.4
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
huggingface-hub==0.10.1
identify==2.5.6
idna==3.4
imageio==2.22.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
install==1.3.5
iopath==0.1.10
ipython==7.34.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
kaggle==1.5.12
kiwisolver==1.4.4
-e git+https://github.com/salesforce/LAVIS.git@c2b5b2f492f546c55e3a1dccd36418628806c5d4#egg=lavis
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
networkx==2.6.3
nodeenv==1.7.0
numpy==1.21.6
omegaconf==2.2.3
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
plotly==4.14.3
portalocker==2.5.1
pre-commit==2.20.0
prompt-toolkit==3.0.31
protobuf==3.20.3
ptyprocess==0.7.0
pyarrow==9.0.0
pycocoevalcap==1.2
pycocotools==2.0.5
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==6.1.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
retrying==1.3.3
rich==12.6.0
scikit-image==0.19.3
scipy==1.7.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.13.0
text-unidecode==1.3
tifffile==2021.11.2
timm==0.4.12
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.10.0+cu102
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu102
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.21.3
transnetv2 @ file:///home/finn/AIC/TransNetV2
typing-extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
virtualenv==20.16.5
watchdog==2.1.9
wcwidth==0.2.5
webdataset==0.2.26
zipp==3.9.0
